Simplify folio client
